backend/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
supabase==1.2.0
Flask-Cors==4.0.0
gunicorn==21.2.0
python-jose==3.3.0
requests==2.31.0
httpx==0.24.1
flask-jwt-extended==4.6.0 